# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * NexusHub Admin API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface SuccessResponse {
  success: boolean;
  message?: string;
}

export interface AdminLoginInput {
  /** @minLength 1 */
  username: string;
  /** @minLength 1 */
  password: string;
}

export type AdminUserRole = typeof AdminUserRole[keyof typeof AdminUserRole];


export const AdminUserRole = {
  super_admin: 'super_admin',
  editor: 'editor',
} as const;

export interface AdminUser {
  id: number;
  username: string;
  email: string;
  role: AdminUserRole;
  totpEnabled: boolean;
  sessionTimeoutMinutes: number;
  createdAt: string;
}

export interface AuthResult {
  token: string;
  user: AdminUser;
}

export interface ChangePasswordInput {
  /** @minLength 1 */
  currentPassword: string;
  /** @minLength 8 */
  newPassword: string;
}

/**
 * @nullable
 */
export type AuditLogOldValue = { [key: string]: unknown } | null;

/**
 * @nullable
 */
export type AuditLogNewValue = { [key: string]: unknown } | null;

export interface AuditLog {
  id: number;
  /** @nullable */
  adminUserId?: number | null;
  adminUsername: string;
  action: string;
  entityType: string;
  /** @nullable */
  entityId?: string | null;
  /** @nullable */
  oldValue?: AuditLogOldValue;
  /** @nullable */
  newValue?: AuditLogNewValue;
  /** @nullable */
  ipAddress?: string | null;
  createdAt: string;
}

export interface DashboardStats {
  totalPlans: number;
  activePlans: number;
  totalPacks: number;
  activePacks: number;
  maintenanceMode?: boolean;
  recentActivity: AuditLog[];
}

export type MembershipPlanBillingCycle = typeof MembershipPlanBillingCycle[keyof typeof MembershipPlanBillingCycle];


export const MembershipPlanBillingCycle = {
  monthly: 'monthly',
  yearly: 'yearly',
  lifetime: 'lifetime',
  one_time: 'one_time',
} as const;

export interface MembershipPlan {
  id: number;
  name: string;
  price: string;
  billingCycle: MembershipPlanBillingCycle;
  description: string;
  features: string[];
  checkoutUrl: string;
  isActive: boolean;
  isFeatured: boolean;
  createdAt: string;
  updatedAt: string;
}

export type MembershipPlanInputBillingCycle = typeof MembershipPlanInputBillingCycle[keyof typeof MembershipPlanInputBillingCycle];


export const MembershipPlanInputBillingCycle = {
  monthly: 'monthly',
  yearly: 'yearly',
  lifetime: 'lifetime',
  one_time: 'one_time',
} as const;

export interface MembershipPlanInput {
  /** @minLength 1 */
  name: string;
  price: string;
  billingCycle: MembershipPlanInputBillingCycle;
  description?: string;
  features?: string[];
  checkoutUrl?: string;
  isActive?: boolean;
  isFeatured?: boolean;
}

export type MembershipPlanUpdateBillingCycle = typeof MembershipPlanUpdateBillingCycle[keyof typeof MembershipPlanUpdateBillingCycle];


export const MembershipPlanUpdateBillingCycle = {
  monthly: 'monthly',
  yearly: 'yearly',
  lifetime: 'lifetime',
  one_time: 'one_time',
} as const;

export interface MembershipPlanUpdate {
  /** @minLength 1 */
  name?: string;
  price?: string;
  billingCycle?: MembershipPlanUpdateBillingCycle;
  description?: string;
  features?: string[];
  checkoutUrl?: string;
  isActive?: boolean;
  isFeatured?: boolean;
}

export interface StarterPack {
  id: number;
  name: string;
  price: string;
  discountPercentage: number;
  items: string[];
  checkoutUrl: string;
  isAvailable: boolean;
  createdAt: string;
  updatedAt: string;
}

export interface StarterPackInput {
  /** @minLength 1 */
  name: string;
  price: string;
  discountPercentage?: number;
  items?: string[];
  checkoutUrl?: string;
  isAvailable?: boolean;
}

export interface StarterPackUpdate {
  /** @minLength 1 */
  name?: string;
  price?: string;
  discountPercentage?: number;
  items?: string[];
  checkoutUrl?: string;
  isAvailable?: boolean;
}

export interface SiteSettings {
  id: number;
  siteName: string;
  /** @nullable */
  logoUrl?: string | null;
  contactEmail: string;
  currency: string;
  taxRate: string;
  /** @nullable */
  notificationEmail?: string | null;
  maintenanceMode: boolean;
  discordInviteUrl: string;
  updatedAt: string;
}

export interface SiteSettingsUpdate {
  /** @minLength 1 */
  siteName?: string;
  /** @nullable */
  logoUrl?: string | null;
  contactEmail?: string;
  currency?: string;
  taxRate?: string;
  /** @nullable */
  notificationEmail?: string | null;
  maintenanceMode?: boolean;
  discordInviteUrl?: string;
}

export interface AuditLogList {
  items: AuditLog[];
  total: number;
}

export type ListAuditLogsParams = {
limit?: number;
offset?: number;
};