# lib/api-zod/src/generated/types/auditLog.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * NexusHub Admin API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AuditLogNewValue } from './auditLogNewValue';
import type { AuditLogOldValue } from './auditLogOldValue';

export interface AuditLog {
  id: number;
  /** @nullable */
  adminUserId?: number | null;
  adminUsername: string;
  action: string;
  entityType: string;
  /** @nullable */
  entityId?: string | null;
  /** @nullable */
  oldValue?: AuditLogOldValue;
  /** @nullable */
  newValue?: AuditLogNewValue;
  /** @nullable */
  ipAddress?: string | null;
  createdAt: string;
}