# lib/api-zod/src/generated/types/auditLogList.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * NexusHub Admin API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AuditLog } from './auditLog';

export interface AuditLogList {
  items: AuditLog[];
  total: number;
}