# lib/api-zod/src/generated/types/membershipPlan.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * NexusHub Admin API specification
 * OpenAPI spec version: 0.1.0
 */
import type { MembershipPlanBillingCycle } from './membershipPlanBillingCycle';

export interface MembershipPlan {
  id: number;
  name: string;
  price: string;
  billingCycle: MembershipPlanBillingCycle;
  description: string;
  features: string[];
  checkoutUrl: string;
  isActive: boolean;
  isFeatured: boolean;
  createdAt: string;
  updatedAt: string;
}