# lib/api-zod/src/generated/types/membershipPlanInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * NexusHub Admin API specification
 * OpenAPI spec version: 0.1.0
 */
import type { MembershipPlanInputBillingCycle } from './membershipPlanInputBillingCycle';

export interface MembershipPlanInput {
  /** @minLength 1 */
  name: string;
  price: string;
  billingCycle: MembershipPlanInputBillingCycle;
  description?: string;
  features?: string[];
  checkoutUrl?: string;
  isActive?: boolean;
  isFeatured?: boolean;
}