# lib/api-zod/src/generated/types/siteSettings.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * NexusHub Admin API specification
 * OpenAPI spec version: 0.1.0
 */

export interface SiteSettings {
  id: number;
  siteName: string;
  /** @nullable */
  logoUrl?: string | null;
  contactEmail: string;
  currency: string;
  taxRate: string;
  /** @nullable */
  notificationEmail?: string | null;
  maintenanceMode: boolean;
  discordInviteUrl: string;
  updatedAt: string;
}