# lib/api-zod/src/generated/types/starterPackUpdate.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * NexusHub Admin API specification
 * OpenAPI spec version: 0.1.0
 */

export interface StarterPackUpdate {
  /** @minLength 1 */
  name?: string;
  price?: string;
  discountPercentage?: number;
  items?: string[];
  checkoutUrl?: string;
  isAvailable?: boolean;
}